Fixed FPS counter bug Fixed Audio play bug
Ticket:MET-47, MET-36
Tests Done: Compiled and verified fps counter and audio is working.

+++ GameEngine/EditorUI.h
@@ -19,7 +19,7 @@ class Editor
 {
 public:
 
-	float last_frame_fps;
+	int last_frame_fps;
 
 	void Init() const;
 	void NewFrame() const;
